requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
pythondialog==3.4.0
requests==2.18.4
urllib3==1.22
